bug fix (see note)
Fixes issue where the tour block fails to validate if the post was not saved at least once before adding the block. The post id is not available when the post has not yet been saved, so there's a discrepancy between the edit and save states. As a workaround, I'm filtering the block to add a wrapper with the post id as a data attribute, which I can use for the same ends (to redirect to the tour single when a user clicks on the view map button).